backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
fpdf2>=2.7.4
Werkzeug==3.0.1
joblib==1.3.2
google-generativeai
python-dotenv
uharfbuzz
groq
